Convert tests to phpunit base classes.
- Fixes #135.
