+++ tests/src/Functional/MessageUiTest.php
@@ -1,17 +1,10 @@
 <?php
 
-/**
- * @file
- * Definition of Drupal\message\Tests\MessageUiTest.
- */
+namespace Drupal\Tests\message\Functional;
 
-namespace Drupal\message\Tests;
-
-use Drupal\Core\Language\Language;
 use Drupal\language\Entity\ConfigurableLanguage;
 use Drupal\message\Entity\Message;
 use Drupal\message\Entity\MessageTemplate;
-use Drupal\user\Entity\User;
 
 /**
  * Testing the CRUD functionality for the Message template entity.
@@ -28,7 +21,7 @@ class MessageUiTest extends MessageTestBase {
   /**
    * The user object.
    *
-   * @var User
+   * @var \Drupal\user\UserInterface
    */
   protected $account;
 
@@ -118,27 +111,6 @@ class MessageUiTest extends MessageTestBase {
     $this->drupalPostForm('admin/structure/message/delete/' . $template, [], 'Delete');
     $this->assertText(t('There is no Message template yet.'));
     $this->assertFalse(MessageTemplate::load($template), 'The message deleted via the UI successfully.');
-  }
-
-  /**
-   * Test that message render returns message text wrapped in a div.
-   */
-  public function testMessageTextWrapper() {
-    $template = 'dummy_message';
-    // Create message to be rendered.
-    $message_template = $this->createMessageTemplate($template, 'Dummy message', '', ['Text to be wrapped by div.']);
-    $message = Message::create(['template' => $message_template->id()])
-      ->setOwner($this->account);
-
-    $message->save();
-
-    // Simulate theming of the message.
-    $build = \Drupal::entityTypeManager()->getViewBuilder('message')->view($message);
-    $output = \Drupal::service('renderer')->renderRoot($build);
-    $this->setRawContent($output);
-    $xpath = $this->xpath('//div');
-
-    $this->assertTrue($xpath, 'A div has been found wrapping the message text.');
   }
 
   /**
