+++ tests/src/Kernel/MessageTemplateCreateTrait.php
@@ -1,14 +1,13 @@
 <?php
 /**
  * @file
- * Contains \Drupal\message\Tests\MessageTemplateCreateTrait.
+ * Contains \Drupal\Tests\message\Kernel\MessageTemplateCreateTrait.
  */
 
-namespace Drupal\message\Tests;
+namespace Drupal\Tests\message\Kernel;
 
 use Drupal\Core\Language\Language;
 use Drupal\message\Entity\MessageTemplate;
-use Drupal\message\MessageTemplateInterface;
 
 /**
  * Trait to assist message template creation for tests.
